packages/lyric_analyzer_base/pyproject.toml
[project]
name = "lyric-analyzer-base"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "google-genai>=1.56.0",
    "lyricsgenius>=3.7.5",
    "pendulum>=3.1.0",
    "polars>=1.36.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
